+++ gcore/gdalrasterband.cpp
@@ -2569,7 +2569,7 @@ GDALRasterBand::SetNoDataValueAsString(const char *pszNoData,
         const float fVal = CPLStrtof(pszNoData, &endptr);
         if (endptr == pszNoData + strlen(pszNoData))
         {
-            return SetNoDataValue(fVal);
+            return SetNoDataValue(double(fVal));
         }
     }
     else
@@ -4385,7 +4385,7 @@ CPLErr GDALRasterBand::GetHistogram(double dfMin, double dfMax, int nBuckets,
                              ARE_REAL_EQUAL(fValue,
                                             sNoDataValues.fNoDataValue)))
                             continue;
-                        dfValue = fValue;
+                        dfValue = double(fValue);
                         break;
                     }
                     case GDT_Float64:
@@ -4429,9 +4429,9 @@ CPLErr GDALRasterBand::GetHistogram(double dfMin, double dfMax, int nBuckets,
                     case GDT_CFloat32:
                     {
                         const double dfReal =
-                            static_cast<float *>(pData)[iOffset * 2];
-                        const double dfImag =
-                            static_cast<float *>(pData)[iOffset * 2 + 1];
+                            double(static_cast<float *>(pData)[iOffset * 2]);
+                        const double dfImag = double(
+                            static_cast<float *>(pData)[iOffset * 2 + 1]);
                         if (std::isnan(dfReal) || std::isnan(dfImag))
                             continue;
                         dfValue = sqrt(dfReal * dfReal + dfImag * dfImag);
@@ -4656,7 +4656,7 @@ CPLErr GDALRasterBand::GetHistogram(double dfMin, double dfMax, int nBuckets,
                                  ARE_REAL_EQUAL(fValue,
                                                 sNoDataValues.fNoDataValue)))
                                 continue;
-                            dfValue = fValue;
+                            dfValue = double(fValue);
                             break;
                         }
                         case GDT_Float64:
@@ -4695,10 +4695,10 @@ CPLErr GDALRasterBand::GetHistogram(double dfMin, double dfMax, int nBuckets,
                         }
                         case GDT_CFloat32:
                         {
-                            double dfReal =
-                                static_cast<float *>(pData)[iOffset * 2];
-                            double dfImag =
-                                static_cast<float *>(pData)[iOffset * 2 + 1];
+                            double dfReal = double(
+                                static_cast<float *>(pData)[iOffset * 2]);
+                            double dfImag = double(
+                                static_cast<float *>(pData)[iOffset * 2 + 1]);
                             if (std::isnan(dfReal) || std::isnan(dfImag))
                                 continue;
                             dfValue = sqrt(dfReal * dfReal + dfImag * dfImag);
@@ -6387,7 +6387,7 @@ static inline double GetPixelValue(GDALDataType eDataType, bool bSignedByte,
                 bValid = false;
                 return 0.0;
             }
-            dfValue = fValue;
+            dfValue = double(fValue);
             return dfValue;
         }
         case GDT_Float64:
@@ -6413,7 +6413,7 @@ static inline double GetPixelValue(GDALDataType eDataType, bool bSignedByte,
             }
             break;
         case GDT_CFloat32:
-            dfValue = static_cast<const float *>(pData)[iOffset * 2];
+            dfValue = double(static_cast<const float *>(pData)[iOffset * 2]);
             if (std::isnan(dfValue))
             {
                 bValid = false;
