+++ ogr/ogrgeojsonwriter.cpp
@@ -1574,7 +1574,7 @@ json_object *
 json_object_new_float_with_significant_figures(float fVal,
                                                int nSignificantFigures)
 {
-    json_object *jso = json_object_new_double(fVal);
+    json_object *jso = json_object_new_double(double(fVal));
     json_object_set_serializer(
         jso, OGR_json_float_with_significant_figures_to_string,
         reinterpret_cast<void *>(static_cast<uintptr_t>(nSignificantFigures)),
